+++ components/atom/label/demo/index.js
@@ -1,4 +1,7 @@
 import AtomLabel, {AtomLabelTypes, AtomLabelFontSizes} from '../src'
+import AtomInput from '@s-ui/react-atom-input'
+import AtomCheckbox from '@s-ui/react-atom-checkbox'
+import AtomButton from '@s-ui/react-atom-button'
 
 import {
   H1,
@@ -8,10 +11,7 @@ import {
   Grid,
   Cell,
   Box,
-  Input,
-  Code,
-  Radio,
-  Button
+  Code
 } from '@s-ui/documentation-library'
 
 const flexCenteredStyle = {
@@ -27,7 +27,7 @@ const Demo = () => {
   const labelTypes = [['default', ''], ...Object.entries(AtomLabelTypes)]
 
   return (
-    <div>
+    <div className="sui-StudioPreview">
       <H1>Label</H1>
       <Paragraph>
         The Label is the name of the associated field, that explains what is the
@@ -46,11 +46,11 @@ const Demo = () => {
                 <AtomLabel
                   name={`atomLabelName-${key}`}
                   for={`labelName-${key}`}
-                  text={`label ${value}`}
-                  optionalText="*"
+                  text={`Label ${value}`}
+                  optionalText="(Optional)"
                   type={value}
                 />
-                <Input />
+                <AtomInput />
               </Box>
             </Cell>
           ))}
@@ -66,9 +66,9 @@ const Demo = () => {
         </Paragraph>
         <Grid cols={3} gutter={[8, 8]}>
           {[
-            <Input key={0} />,
-            <Radio key={1} />,
-            <Button key={2}>Button</Button>
+            <AtomInput key={0} />,
+            <AtomCheckbox key={1} />,
+            <AtomButton key={2}>Button</AtomButton>
           ].map((component, index) =>
             ['left', undefined, 'right'].map((value, index) => (
               <Cell key={index} style={flexCenteredStyle}>
@@ -77,8 +77,8 @@ const Demo = () => {
                   <AtomLabel
                     name={`atomLabelName-${value}`}
                     for={`labelName-${value}`}
-                    text={`label ${value}`}
-                    optionalText="*"
+                    text={`Label ${value}`}
+                    optionalText="(Optional)"
                     inline={value}
                   />
                   {value !== 'right' && component}
@@ -103,10 +103,10 @@ const Demo = () => {
                   <AtomLabel
                     name={`atomLabelName-${key}`}
                     for={`labelName-${key}`}
-                    text={`size ${value}`}
+                    text={`Size ${value}`}
                     fontSize={value}
                   />
-                  <Input />
+                  <AtomInput />
                 </Box>
               </Cell>
             )