+++ components/atom/table/src/settings.js
@@ -1,0 +1,16 @@
+export const BASE_CLASS = 'react-AtomTable'
+export const CELL_BASE_CLASS = `${BASE_CLASS}-cell`
+export const ROW_BASE_CLASS = `${BASE_CLASS}-row`
+
+export const TABLE_CELL_TYPE = {
+  header: 'th',
+  data: 'td'
+}
+
+export const TABLE_CELL_PADDING = {
+  xs: 'xs',
+  s: 's',
+  m: 'm',
+  l: 'l',
+  xl: 'l'
+}