+++ components/molecule/dropdownList/test/index.test.js
@@ -55,6 +55,21 @@ describe('molecule/dropdownList', () => {
     expect(container.innerHTML).to.not.have.lengthOf(0)
   })
 
+  it('should NOT extend classNames', () => {
+    // Given
+    const props = {
+      className: 'extended-classNames'
+    }
+    const findSentence = str => string => string.match(new RegExp(`S*${str}S*`))
+
+    // When
+    const {container} = setup(props)
+    const findClassName = findSentence(props.className)
+
+    // Then
+    expect(findClassName(container.innerHTML)).to.be.null
+  })
+
   it('should forward a given ref to the `ul` element', () => {
     // Given
     const props = {}