+++ components/molecule/dropdownOption/src/index.js
@@ -25,6 +25,7 @@ const MoleculeDropdownOption = ({
   children,
   disabled,
   highlightQuery,
+  highlightValue,
   innerRef,
   onSelectKey,
   onSelect,
@@ -73,12 +74,16 @@ const MoleculeDropdownOption = ({
       )}">`,
       endTag: '</mark>'
     })
+
     return (
-      <span
-        onFocus={handleInnerFocus}
-        dangerouslySetInnerHTML={{__html: mark}}
-        className={innerClassName}
-      />
+      <>
+        <span
+          onFocus={handleInnerFocus}
+          dangerouslySetInnerHTML={{__html: mark}}
+          className={innerClassName}
+        />
+        {highlightValue ? children : null}
+      </>
     )
   }
   const handleInnerFocus = ev => {
@@ -102,7 +107,7 @@ const MoleculeDropdownOption = ({
         />
       )}
       {highlightQuery ? (
-        renderHighlightOption(children)
+        renderHighlightOption(highlightValue || children)
       ) : (
         <span onFocus={handleInnerFocus} className={innerClassName}>
           {children}
@@ -131,6 +136,8 @@ MoleculeDropdownOption.propTypes = {
   selected: PropTypes.bool,
   /** Text to be highlighted in the option text if found */
   highlightQuery: PropTypes.string,
+  /** Text to be display if used with highlight query with custom content */
+  highlightValue: PropTypes.string,
   /* key to provoke the onClick callback. Valid any value defined here → https://www.w3.org/TR/uievents-key/#named-key-attribute-values */
   onSelectKey: PropTypes.oneOfType([PropTypes.string, PropTypes.array]),
   /** Custom ref handler that will be assigned to the "target" element */