+++ components/molecule/inputTags/test/index.test.js
@@ -12,50 +12,115 @@ describe('molecule/inputTags', () => {
   const Component = MoleculeInputTags
   const setup = setupEnvironment(Component)
 
+  it('should render without crashing', () => {
+    // Given
+    const props = {}
+
+    // When
+    const component = <Component {...props} />
+
+    // Then
+    const div = document.createElement('div')
+    ReactDOM.render(component, div)
+    ReactDOM.unmountComponentAtNode(div)
+  })
+
+  it('should NOT render null', () => {
+    // Given
+    const props = {}
+
+    // When
+    const {container} = setup(props)
+
+    // Then
+    expect(container.innerHTML).to.be.a('string')
+    expect(container.innerHTML).to.not.have.lengthOf(0)
+  })
+
+  it('should NOT extend classNames', () => {
+    // Given
+    const props = {
+      className: 'extended-classNames'
+    }
+    const findSentence = str => string => string.match(new RegExp(`S*${str}S*`))
+
+    // When
+    const {container} = setup(props)
+    const findClassName = findSentence(props.className)
+
+    // Then
+    expect(findClassName(container.innerHTML)).to.be.null
+  })
+
   describe('when the is disabled', () => {
     it('should be disabled', () => {
-      const {queryByRole} = setup({disabled: true})
+      // Given
+      const props = {disabled: true}
 
+      // When
+      const {queryByRole} = setup(props)
       const input = queryByRole('textbox')
+
+      // Then
       expect(input).to.be.null
     })
   })
 
   describe('when has placeholder', () => {
     it('should display the placeholder if no tags avaiable', () => {
-      const placeholder = 'Type your favorite beetle'
-      const {getByPlaceholderText} = setup({placeholder})
+      // Given
+      const props = {placeholder: 'Type your favorite beetle'}
 
-      expect(getByPlaceholderText(placeholder)).to.be.visible
+      // When
+      const {getByPlaceholderText} = setup(props)
+      const expected = getByPlaceholderText(props.placeholder)
+
+      // Then
+      expect(expected).to.be.visible
     })
 
     it('should not display the placeholder after adding tags', async () => {
-      const placeholder = 'Type your favorite beetle'
+      // Given
+      const props = {placeholder: 'Type your favorite beetle', tags: ['Lenon']}
 
-      const {queryByPlaceholderText} = setup({
-        placeholder,
-        tags: ['Lenon']
-      })
+      // When
+      const {queryByPlaceholderText} = setup(props)
+      const expected = queryByPlaceholderText(props.placeholder)
 
-      expect(queryByPlaceholderText(placeholder)).to.be.null
+      // then
+      expect(expected).to.be.null
     })
   })
 
   describe('when has maxTags', () => {
     it('should allow add tags if max not reached', () => {
-      const {getByRole} = setup({
+      // Given
+      const props = {
         maxTags: 4,
         tags: []
-      })
-      expect(getByRole('textbox')).to.be.visible
+      }
+
+      // When
+      const {getByRole} = setup(props)
+      const expected = getByRole('textbox')
+
+      // Then
+      expect(expected).to.be.visible
     })
 
     it('should not allow add tags if max reached', () => {
-      const {queryByRole} = setup({
+      // Given
+      const props = {
         maxTags: 4,
         tags: ['Paul', 'John', 'Ringo', 'George']
-      })
-      expect(queryByRole('textbox')).to.be.null
+      }
+
+      // When
+      const {queryByRole} = setup(props)
+      const expected = queryByRole('textbox')
+
+      // Then
+      expect(expected).to.be.null
     })
   })
 })