+++ components/molecule/modal/demo/ArticleSize.js
@@ -7,7 +7,6 @@ import {
   RadioButton,
   RadioButtonGroup,
   Button,
-  Label,
   Grid,
   Cell
 } from '@s-ui/documentation-library'
@@ -21,8 +20,8 @@ const ArticleSize = ({className}) => {
   const [count, setCount] = useState(3)
   const [fitContent, setFitContent] = useState(false)
 
-  const onChangeHandler = (_, value) => {
-    setOpen(value === undefined ? open : value)
+  const onChangeHandler = () => {
+    setOpen(!open)
   }
 
   const onCloseButtonHandler = () => {
@@ -53,17 +52,8 @@ const ArticleSize = ({className}) => {
   return (
     <Article className={className}>
       <H2>Sizes</H2>
-      <Grid cols={1} gutter={[8, 8]}>
-        <Cell>
-          <Label>isOpen</Label>
-        </Cell>
-        <Cell>
-          <RadioButtonGroup value={open} onChange={onChangeHandler}>
-            <RadioButton value label="true" checked={open === true} />
-            <RadioButton value={false} label="false" checked={open === false} />
-          </RadioButtonGroup>
-        </Cell>
-      </Grid>
+      <Button onClick={onChangeHandler}>Open modal</Button>
+
       <Paragraph>
         Sizes can be customized under the <Code>size</Code> prop.
       </Paragraph>