+++ components/atom/label/CHANGELOG.md
@@ -1,5 +1,25 @@
 # CHANGELOG
 
+# 1.17.0 (2021-10-16)
+
+
+
+# 1.1.0 (2021-10-16)
+
+
+### Bug Fixes
+
+* **components/atom/label/demo:** hotfix unclosed JSX tag ([ad58f8c](https://github.com/SUI-Components/sui-components/commit/ad58f8c785bb05fbf3478c8a1f6e3abb977c1e77))
+* **components/atom/label:** Changed demo label and src ([78e0961](https://github.com/SUI-Components/sui-components/commit/78e0961b19120d0ae8daf4e042a43b790e50da84))
+
+
+### Features
+
+* **components/atom/label:** change optional message and text input ([a96032a](https://github.com/SUI-Components/sui-components/commit/a96032a0130e576fddab7a3254dd44d339ff709c))
+* **components/atom/label:** deleted duplicate import ([e34eab1](https://github.com/SUI-Components/sui-components/commit/e34eab1e98917b7104ef9e68f916f63bd059c6ca))
+
+
+
 # 1.16.0 (2021-09-09)